+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added: methods to handle events are underscored when String#underscore is
   defined, e.g. `def MyEvent` versus `def my_event`.
+- Added: Publisher#unsubscribe method
 
 ## [0.1.0]